requirements.txt
aiosqlite==0.22.1
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.7
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==25.4.0
beartype==0.22.9
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==3.0.0
croniter==6.0.0
cryptography==46.0.4
dateparser==1.2.2
docker==7.1.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.128.0
fsspec==2026.1.0
graphviz==0.21
greenlet==3.3.1
griffe==1.15.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lupa==2.6
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.1
oauthlib==3.3.1
opentelemetry-api==1.39.1
orjson==3.11.6
outcome==1.3.0.post0
packaging==25.0
pandas==3.0.0
pathspec==1.0.4
pluggy==1.6.0
prefect==3.6.15
prometheus_client==0.24.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.17.4
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
redis==7.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.1
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
scikit-learn==1.8.0
scipy==1.17.0
selenium==4.39.0
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
starlette==0.50.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
trio==0.32.0
trio-websocket==0.12.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==16.0
whenever==0.9.5
wsproto==1.3.2
xgboost==3.1.3
zipp==3.23.0
